pocketBackend/project/requirements.txt
asgiref==3.8.1
cachetools==6.2.4
Django==5.1.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.45.0
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
pillow==12.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.2.1
rsa==4.9.1
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
